Move `VisualShader` to its own folder, allow to build export template without it
