+++ editor/inspector/editor_properties.cpp
@@ -57,7 +57,7 @@
 #include "scene/main/window.h"
 #include "scene/resources/font.h"
 #include "scene/resources/mesh.h"
-#include "scene/resources/visual_shader_nodes.h"
+#include "scene/resources/visual_shader/visual_shader_nodes.h"
 
 ///////////////////// NIL /////////////////////////
