+++ editor/shader/visual_shader_editor_plugin.cpp
@@ -68,8 +68,8 @@
 #include "scene/main/window.h"
 #include "scene/resources/curve_texture.h"
 #include "scene/resources/style_box_flat.h"
-#include "scene/resources/visual_shader_nodes.h"
-#include "scene/resources/visual_shader_particle_nodes.h"
+#include "scene/resources/visual_shader/visual_shader_nodes.h"
+#include "scene/resources/visual_shader/visual_shader_particle_nodes.h"
 #include "servers/display_server.h"
 #include "servers/rendering/shader_preprocessor.h"
 #include "servers/rendering/shader_types.h"
