+++ scene/resources/visual_shader/visual_shader_nodes.cpp
