+++ scene/resources/visual_shader/visual_shader_nodes.h
@@ -31,7 +31,7 @@
 #pragma once
 
 #include "scene/resources/curve_texture.h"
-#include "scene/resources/visual_shader.h"
+#include "scene/resources/visual_shader/visual_shader.h"
 
 class Cubemap;
 class Texture2DArray;
