+++ scene/resources/visual_shader/visual_shader_sdf_nodes.h
@@ -30,7 +30,7 @@
 
 #pragma once
 
-#include "scene/resources/visual_shader.h"
+#include "scene/resources/visual_shader/visual_shader.h"
 
 class VisualShaderNodeSDFToScreenUV : public VisualShaderNode {
 	GDCLASS(VisualShaderNodeSDFToScreenUV, VisualShaderNode);
